WebIDL/tests/submissions/heycam/Date-001.html

<!DOCTYPE html>
<title>Date</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.26 Date
//
// Converting any JS value other than a Date object will throw a TypeError.
// Returning a Date object from an attribute will create a new object
// every time.

// feature: HTMLInputElement.valueAsDate (attribute) (of type Date)

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  input.value = "2011-01-01";
  var date = input.valueAsDate;
  assert_true(date instanceof Date);
  assert_equals(date.getYear(), 111);
}, "test getting HTMLInputElement.valueAsDate");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  input.value = "2011-01-01";
  var date1 = input.valueAsDate;
  var date2 = input.valueAsDate;
  assert_true(date1 !== date2);
}, "test getting HTMLInputElement.valueAsDate twice");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  input.value = "";
  var date = input.valueAsDate;
  assert_equals(+date, NaN);
}, "test getting HTMLInputElement.valueAsDate when it represents an undefined time, after assigning the empty string");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  input.valueAsDate = null;
  assert_equals(+input.valueAsDate, NaN);
}, "test getting HTMLInputElement.valueAsDate when it represents an undefined time, after assigning null");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  input.valueAsDate = new Date("2011-01-01");
  assert_true(input.value.indexOf("2011-01-01") == 0);
}, "test assigning a Date object to HTMLInputElement.valueAsDate");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  assert_throws(new TypeError(), function() { input.valueAsDate = Object.create(new Date()); });
}, "test assigning { } -> Date to HTMLInputElement.valueAsDate");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  assert_throws(new TypeError(), function() { input.valueAsDate = 123; });
}, "test assigning 123 to HTMLInputElement.valueAsDate");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  assert_throws(new TypeError(), function() { input.valueAsDate = "2011-01-01"; });
}, "test assigning \"2011-01-01\" to HTMLInputElement.valueAsDate");

test(function() {
  var input = document.createElement("input");
  input.type = "date";
  var date = new Date("2011-01-01");
  date.valueOf = function() { throw "This exception should not have been thrown"; };
  input.valueAsDate = date;
  assert_equals(input.valueAsDate.getYear(), "111");
}, "test assigning a Date object with a throwing valueOf method to HTMLInputElement.valueAsDate");
</script>
